begin implementation of graphql api for contacts
