+++ source/cosmic.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{DE7AE29D-A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cosmic.Tests.Shared", "test\Cosmic.Tests.Shared\Cosmic.Tests.Shared.csproj", "{A582494A-62D3-46F6-B615-2C6AA0113025}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cosmic.Server.Core", "src\Cosmic.Server.Core\Cosmic.Server.Core.csproj", "{E873B747-EEA5-478D-A603-6FBAF4D21826}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -25,6 +27,10 @@ Global
 		{A582494A-62D3-46F6-B615-2C6AA0113025}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A582494A-62D3-46F6-B615-2C6AA0113025}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A582494A-62D3-46F6-B615-2C6AA0113025}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E873B747-EEA5-478D-A603-6FBAF4D21826}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E873B747-EEA5-478D-A603-6FBAF4D21826}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E873B747-EEA5-478D-A603-6FBAF4D21826}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E873B747-EEA5-478D-A603-6FBAF4D21826}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -32,6 +38,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{2CAA63D6-1869-4AE5-A2D5-F3C2C4E0D01B} = {B48B50F0-AA81-4EEF-80C4-6A765A8B3B34}
 		{A582494A-62D3-46F6-B615-2C6AA0113025} = {DE7AE29D-A8FF-4862-BA34-D25FD60978C6}
+		{E873B747-EEA5-478D-A603-6FBAF4D21826} = {B48B50F0-AA81-4EEF-80C4-6A765A8B3B34}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E9CF8D81-1F94-4062-AF59-826812420299}
