+++ source/src/Cosmic.Server.Core/Cosmic.Server.Core.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Cosmic.Shared\Cosmic.Shared.csproj" />
+  </ItemGroup>
+
+</Project>
